feat(plugin-assistant): add RoutineProperties story and split AgentProperties into component + container
- Add components/RoutineProperties: presentational component wrapping
  ObjectProperties for a Routine, plus a withClientProvider-based story.
- Push AgentProperties UI down to components/AgentProperties (no
  @dxos/app-framework or @dxos/app-toolkit deps); keep a thin container
  wrapper that owns the SyncTriggers side effect via useSpaceCallback.
- Move AgentProperties story to components with the same simplified
  decorator stack (withTheme + withLayout + withClientProvider), replacing
  the prior plugin-manager-based story.
- Fix react-ui-form: FormInputAnnotation was being stripped from non-keyword
  inner types (Struct/Array) by encodedBoundAST inside getProperties, so
  Routine.input/output/instructions/context still rendered. getFormProperties
  now reads FormInputAnnotation from the raw property signatures (mirroring
  omitHiddenFormFields) and filters the encoded result by name. Adds
  regression tests covering keyword, struct, array, and chained-.annotations()
  cases.
