+++ packages/plugins/plugin-assistant/src/components/AgentProperties/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './AgentProperties';
