+++ packages/plugins/plugin-assistant/src/components/RoutineProperties/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export { RoutineProperties as default } from './RoutineProperties';
