+++ packages/plugins/plugin-assistant/src/containers/AgentProperties/AgentProperties.tsx
@@ -2,32 +2,23 @@
 // Copyright 2025 DXOS.org
 //
 
-import { Atom } from '@effect-atom/atom';
-import { useAtomValue } from '@effect-atom/atom-react';
-import * as Option from 'effect/Option';
-import React, { useCallback, useEffect, useMemo } from 'react';
+import React, { useEffect } from 'react';
 
 import { useSpaceCallback } from '@dxos/app-framework/ui';
 import { type AppSurface } from '@dxos/app-toolkit/ui';
 import { Agent, SyncTriggers } from '@dxos/assistant-toolkit';
 import { Operation } from '@dxos/compute';
-import { DXN, Filter, Obj, Ref } from '@dxos/echo';
-import { AtomObj } from '@dxos/echo-atom';
+import { Obj, Ref } from '@dxos/echo';
 import { log } from '@dxos/log';
-import { useQuery } from '@dxos/react-client/echo';
-import { Input, useTranslation } from '@dxos/react-ui';
-import { FeedAnnotation } from '@dxos/schema';
 
-import { meta } from '#meta';
+import { AgentProperties as AgentPropertiesComponent } from '#components';
 
 export type AgentPropertiesProps = AppSurface.ObjectPropertiesProps<Agent.Agent>;
 
 export const AgentProperties = ({ subject: agent }: AgentPropertiesProps) => {
-  const { t } = useTranslation(meta.id);
-  const db = Obj.getDatabase(agent);
-
   const spaceId = Obj.getDatabase(agent)?.spaceId;
 
+  // TODO(burdon): Factor this out?
   const syncTriggers = useSpaceCallback(
     spaceId,
     [] as const,
@@ -35,8 +26,9 @@ export const AgentProperties = ({ subject: agent }: AgentPropertiesProps) => {
     [agent],
   );
 
+  // Sync triggers whenever the agent mutates (e.g., subscriptions change).
   useEffect(() => {
-    if (!db) {
+    if (!spaceId) {
       return;
     }
 
@@ -45,82 +37,7 @@ export const AgentProperties = ({ subject: agent }: AgentPropertiesProps) => {
         syncTriggers().catch((err) => log.catch(err));
       });
     });
-  }, [agent, syncTriggers]);
+  }, [spaceId, agent, syncTriggers]);
 
-  // Build a filter matching objects of any schema annotated as a feed.
-  const feedFilter = useMemo(() => {
-    if (!db) {
-      return Filter.nothing();
-    }
-
-    const schemas = db.schemaRegistry.query({ location: ['database', 'runtime'] }).runSync();
-    const feedSchemas = schemas.filter((schema) => {
-      const annotation = FeedAnnotation.get(schema);
-      return Option.isSome(annotation) && annotation.value === true;
-    });
-
-    return feedSchemas.length === 0 ? Filter.nothing() : Filter.or(...feedSchemas.map((schema) => Filter.type(schema)));
-  }, [db]);
-
-  const subscribedObjects = useQuery(db, feedFilter);
-
-  // Query all existing subscriptions (e.g., mail, calendar, etc.)
-  const existingSubscriptions = useAtomValue(
-    useMemo(
-      () =>
-        AtomObj.make(agent).pipe((_) =>
-          Atom.make((get) => {
-            const agentObj = get(_);
-            const selectedSubscriptions: Obj.Unknown[] = subscribedObjects.filter((object) =>
-              agentObj.subscriptions.some((subscription) => DXN.equals(subscription.dxn, Obj.getDXN(object))),
-            );
-
-            return selectedSubscriptions;
-          }),
-        ),
-      [agent, subscribedObjects],
-    ),
-  );
-
-  // Create/remove agent subscription.
-  const handleSubscriptionChange = useCallback(
-    (object: Obj.Unknown, checked: boolean) => {
-      Obj.update(agent, (agent) => {
-        if (checked) {
-          agent.subscriptions.push(Ref.fromDXN(Obj.getDXN(object)));
-        } else {
-          agent.subscriptions = agent.subscriptions.filter(
-            (subscription) => !DXN.equals(subscription.dxn, Obj.getDXN(object)),
-          );
-        }
-      });
-    },
-    [agent],
-  );
-
-  if (subscribedObjects.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className='dx-expander flex flex-col'>
-      <Input.Root>
-        <Input.Label classNames='mt-form-gap'>{t('subscriptions.label')}</Input.Label>
-      </Input.Root>
-
-      {subscribedObjects.map((object) => (
-        <Input.Root key={object.id}>
-          <div className='flex items-center gap-2'>
-            <Input.Checkbox
-              checked={existingSubscriptions.includes(object)}
-              onCheckedChange={(checked) => {
-                handleSubscriptionChange(object, checked === true);
-              }}
-            />
-            <Input.Label>{Obj.getLabel(object) ?? object.id}</Input.Label>
-          </div>
-        </Input.Root>
-      ))}
-    </div>
-  );
+  return <AgentPropertiesComponent agent={agent} />;
 };
